barbary_backend/requirements.txt
asgiref==3.5.0
Django==3.2
django-cors-headers==3.11.0
django-jalali==5.1.0
django-jalali-date==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
jdatetime==4.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
